Converted to WorkerComponent
Removed the ScriptComponent, and replaced it with a WorkerComponent. And replaced Start/SyncScript with IStart, and IUpdate respectively.

It works mostly, I think but freezes the editor. I think this is because scripts in the projects will use ScriptComponent. If I find the class that handles auto upgrading scripts. I think that would fix it and it would work.